% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smlx-init.R
\name{initRssimulx}
\alias{initRssimulx}
\title{Initialize Rssimulx library}
\usage{
initRssimulx(software = "simulx", errors = TRUE, warnings = TRUE, info = FALSE)
}
\arguments{
\item{software}{(\emph{optional}) in {"simulx", "monolix"} The software used by R connectors (default 'simulx')}

\item{errors}{(\emph{optional}) (\emph{boolean}) Display/Hide lixoft connectors errors (default TRUE)}

\item{warnings}{(\emph{optional}) (\emph{boolean}) Display/Hide lixoft connectors warnings (default TRUE)}

\item{info}{(\emph{optional}) (\emph{boolean}) Display/Hide lixoft connectors info (default FALSE)}
}
\value{
A boolean equaling TRUE if the initialization has been successful and FALSE if not.
}
\description{
Initialize Rssimulx library
}
\examples{
\dontrun{
initRssimulx()
}
}
