% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smlx-tools.R
\name{inlineDataFrame}
\alias{inlineDataFrame}
\title{Inline dataframe}
\usage{
inlineDataFrame(str)
}
\arguments{
\item{str}{(\emph{string}) Dataframe in string format}
}
\value{
dataframe object
}
\description{
Convert a string in dataframe and save it in a temporary file
}
\examples{
\dontrun{
occ <- inlineDataFrame("
 id time occ
 1   0    1
 1   12   2
 1   24   3
 2   0    1
 2   24   2
 3   0    1
")
}
}
