% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lixoft.read.table}
\alias{lixoft.read.table}
\title{Read Lixoft@ files}
\usage{
lixoft.read.table(file, sep = "", ...)
}
\arguments{
\item{file}{file path of the file to read}

\item{sep}{separator}

\item{\dots}{see \code{\link{read.table}}}
}
\value{
a dataframe object
}
\description{
Utility function to read Lixoft@ formated input/output files
}
