% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsSimulx-deprecated.R
\name{RsSimulx-deprecated}
\alias{RsSimulx-deprecated}
\title{Deprecated functions in package \pkg{RsSimulx}.}
\description{
The functions listed below are deprecated and have been removed
from the package. Alternative functions with similar
functionality are given.
}
\section{\code{writeData}}{

For \code{writeData}, use \code{exportSimulatedData()} in the lixoftConnectors package.
}

\keyword{internal}
