% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rssimulxDemo.model}
\alias{rssimulxDemo.model}
\title{Simulx project}
\format{
A vector of string
}
\source{
Simulx model
}
\usage{
rssimulxDemo.model
}
\description{
Model definition corresponding to rssimulxDemo.smlx project
}
\keyword{datasets}
