% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rssimulxDemo.project}
\alias{rssimulxDemo.project}
\title{Simulx project}
\format{
A vector of string
}
\source{
Simulx project
}
\usage{
rssimulxDemo.project
}
\description{
rssimulxDemo.smlx is a Simulx project.
In this demo three groups with different dose levels are simulated: low, medium and high.
Groups have the same number of individuals, population parameters, distribution of covariates and outputs.
}
\keyword{datasets}
