% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiManager.R
\name{initRsSimulx}
\alias{initRsSimulx}
\title{Initialize RsSimulx library}
\usage{
initRsSimulx(path = NULL, ...)
}
\arguments{
\item{path}{(\emph{character}) (\emph{optional}) Path to installation directory of the Lixoft suite.
If RsSimulx library is not already loaded and no path is given, the directory written in the lixoft.ini file is used for initialization.}

\item{...}{(\emph{optional}) Extra arguments passed to lixoftConnectors package when RsSimulx is used with a version of Lixoft(/@) software suite.
\itemize{
\item \code{force} (\emph{bool}) (\emph{optional}) Should RsSimulx initialization overpass lixoftConnectors software switch security or not. Equals FALSE by default.
}}
}
\value{
A list:
\itemize{
\item \code{software}: the software that is used (should be monolix with Rsmlx)
\item \code{path}: the path to MonolixSuite
\item \code{version}: the version of MonolixSuite that is used
\item \code{status}: boolean equaling TRUE if the initialization has been successful.
}
}
\description{
Initialize RsSimulx library
}
\examples{
\dontrun{
initRsSimulx(path = "/path/to/lixoftRuntime/")
}
}
